Php-Fpm: snapshot data does not show up
snapshot data was previously not showing because it wasn't filled in the
backend. I changed the backend to send flat keys insteaf of nested keys
now. This requires to change how we access the snapshot data in the
ui-client. Data is now showing.

Also applying some changes @bripkens requested in the last PR.