Php-Fpm: wrong date for start time
because the fpm sensor will report start time as a unix timestam, e.g.
seconds and formatDateTime expects miliseconds, the date shown in the
dashboard was wrong. While it makes sense to have the sensor report in
miliseconds already, doing that change here is cheaper due to the
changes needed in agent-update-site, sensors and forge. Also, since the
original value is a unix timestamp, turning it to miliseconds will not
yield any more information.