Support document position based viewport stickyness
This adds support for new simulated positioning type similar to
position:sticky and position:fixed. In contrast to these two though, we
first identify the element's position relative to the document and then
position it in a way so that it is fixed to the viewport at that very
same location.

In addition to this, it also ensures that the content maintains its
original position. In a best case, this repositioning does not result in
a visual reflow.