Add a component to visualize progress
This component is a placeholder for now. It accept our progress objects
and turns this into something visible. It does not currently generating
anything pleasingly looking, as this depends on design and I didn't want
to build anything for the bin.

It implements the <delay before showing progress indicator> pattern to
improve the perceived performance. This delay is by default set to
500ms, but configurable via a property.

I also upgraded storybook because of some funky knobs visible/invisible
bugs.