Do not expose technical messages to users
The backend will now respond with more technical error messages. These
do not really make sense to users. Instead of showing them to users, we
should log them. This will forward them to us via ui-tracker and give us
additional insights into possible reasons for these errors.

This is the counterpart to the following PR, but can be merged
independently:

https://github.com/instana/backend/pull/1414