Avoid flickering breadcrumb
I tried to to this via a new HOC. Turns out that the breadcrumb was
always unmounted and therefore all previous component state was lost.
This made this solution obsolete. Nevertheless, I am pushing the new HOC
as it will come in useful in other cases.

The simple fix was to load the application by ID alone…