Cursor based pagination HOC
Establish a higher-order component for cursor based pagination. The HOC
takes care of retrieving data, merging the various loading steps and
selecting the fitting cursor. It also takes care of resetting the data
when certain properties change.

The component is unit tested with Enzyme [1]. The component will most
likely be used within the analyze context in order to retrieve trace
groups and the trace list.

[1] http://airbnb.io/enzyme/