set v2 parameter on page load and persist user setting it on switch
This also adds a preliminary version of the VersionSwitcher component
(aka magic button), just to be able to check the v2Enabled store's
functionality. We probably need a design for that thing and there is
a dedicated Pivotal task for the magic button, in which we will replace
it with the correct design.