Expose a CSS identifier registry
We are using CSS modules with local names. This results in CSS
identifiers being generated by Webpack. In our current configuration,
this means that we have readable names in the development mode, e.g.
`in-applications_lists_ServicesList_mless___filter` and hashes of that
name in production builds.

This is very nice for development, but it makes automatic UI based
testing complicated. This is because with tests such as Selenium tests,
(xpath|css) selectors need to be written which match the elements with
which should be interacted or clicked. Of course, matching hashes makes
for unreadable UI tests, because something like
`click('.189dsaghjk3219')`
doesn't any sense to a human.

In order to allow us to use these auto generated hashes in production
builds (they provide code obfuscation and remove the chance of name
clashes) and to allow QA to write meaningful UI tests, this PR
adds the generation of a CSS identifier registration file to the build
process.

The file is automatically generated during the Webpack build via a
Webpack plugin. The file has the following structure and is exposed via
the path `/bundle/cssIdentifiers.json`.

```
{
  "in-applications_lists_ApplicationsList_mless___block": "i_O",
  "in-applications_lists_ApplicationsList_mless___flexWrapper": "i_N",
  "in-applications_lists_ApplicationsList_mless___entityIcon": "i_j",

"in-applications_lists_ApplicationsList_mless___noApplicationsDefined":
"i_F",
  "in-applications_lists_ServicesList_mless___block": "i_Q",
  "in-applications_lists_ServicesList_mless___filter": "i_X",
  …
}
```

The file maps from a human readable and stable name to the automatically
generated CSS identifier (CSS class name). With the help of this file,
tests can be written that are meaningful to QA. That should make
@doertedev happy! :)