Turn 2.0 preview mode into proper 2.0 mode
This reactivates a whole bunch of 1.0 features within the 2.0 mode. 2.0
mode therefore turns into a full monitoring solution again. Parts that
we will touch as part of the 2.0 integration work, e.g. events, timeline
and custom issues, were not reactivated.

We will consider how to reactivate these features and what in what form
as part of the respective stories for these items.

This PR also removes the very old and obsolete EUM config view (which
was replaced in mid 2017 with configuration views inside the website
monitoring area).