do not attempt to render trace button row when id is missing
This fixes the "Subscription payload is invalid" on the trace button
on explore dashboards. The problem was probably introduced in
Propbably intro with 833f40999996f7bbcf05659b89c33975d1f2335f
as this removed the check for enitityId in
TraceButtonDialogPresenter/Row.es6

Now we simply do not render the Row at all if the required ID is not
present.