Hierarchical links do not respect time configs (#1109)
Hierarchical links for the 1.0 / 2.0 trace views do not respect time
configs when these are defined. This results in hierarchies to not be
available when the entity isn't available at the focused moment anymore.

This PR extends all the functions and components inside the
`HierarchicalLink` to respect a `timeConfig` property. Additionally,
`HierarchicalLink` will decide whether a change of time config is
necessary. This was done so that the time config changes do not on every
navigation. As this would be insanely confusing…