Allow open in new tab for clicks on groups in analyze
This is currently not possible because there is an `onClick` listener on
each group. In order to fix this, we need to add an `href` to each link.
The problem here is, that the generation of the URL / the mutation of
the URL is done within the `withUrlDependingState` HOC.

This PR extends the `withUrlDependingState` HOC in such a way, that it
exposes an additional reducer which will not call `mutateUrl`, but
instead `getModifiedUrlStream`. This allows generation of links without
losing encapsulation.