Automatically scroll / highlighted selection beacon in page load view
When analyzing errors, resources or HTTP requests, users are selecting
very specific items, but the page load view isn't highlighting these
items. With these changes, the selected items will be scrolled into the
viewport and are temporarily highlighted with a yellow background color.

In order to avoid any more one-offs for functionality like this, this
changes introduces a reusable component that ensures that this is
working properly. On the presentation side, a less mixin exists for
consistent presentation style.

Implements: https://www.pivotaltracker.com/story/show/163014001