Quick filter changes filter it should not change (#1226)
* Quick filter changes filter it should not change

Suppose you add a filter in the website monitoring analyze area via the
quick filter bar, e.g. you filter for a specific browser. This is turned
into a filter with an `EQUALS` operator. This is correct behavior.

Unfortunately when populating the dialog, bar and when manipulating the
list of filters we didn't always respect the operator type. This results
in problems when doing the following:

 1. Open up page view analysis
 2. Add a filter `beacon.browser.name DOES NOT equal Chrome`
 3. The bar item will select Chrome.

This PR fixes the bar behavior to correctly respect the operator type.

* Update packages/in-analyze/tagFiltersHoc.es6

Co-Authored-By: bripkens <bripkens@gmail.com>

* Update packages/in-analyze/tagFiltersHoc.es6

Co-Authored-By: bripkens <bripkens@gmail.com>