Reset pinned metrics when navigating away from dashboard
Correctly reset the pinned metrics when navigating away from a dashboard
that is using the `CustomMetricsV2` components.

This is now easily possible thanks to the new `withUrlState` HOC.

The list of pinned metrics isn't getting reset when navigating between
dashboards. This means that the URL will continue to grow as long as one
never the leaves the dashboard.