Basic Custom Event Support
**Why**

ui-client is already using the internal-only website monitoring custom
event API. It would be great if that reported data would already be
accessible via an analyze data source.

This will enable us to potentially get rid of the ui-tracker earlier. It
will also make the website monitoring data a lot more interesting for
us.

**What**

Adds a new datasource for custom events and ensures proper presentation
within the analyze area.

This is internal only for now. This is also the reason why there is no
dedicated icon for custom events. The proper implementation will be done
in a few months.