Sticky side bar based on position: fixed and position: absolute
Most third party libs I checked (like
https://github.com/abouolia/sticky-sidebar) either require JQuery
and/or do some fancy offset calculation and either use position: top
and change the top value on each scroll event or use
transform: translate[3d] and move the sidebar around with that on each
scroll event.

This is actually not necessary. All we need to do is watch out for
certain events (sidebar reaches viewport top/bottom) and set the
appropriate position attributet push for the given situation.