duplicate old pages
* Alerting Configurations now become Event Filters
* Alerting Integration now become Alert Channels

Both concepts stay more or less unchanged with the events & alerting
redesign, but the names have changed. There will also be further steps
which change the layout of both detail pages later.

Since we need to keep the changes behind a feature flag (thus, keep
both the old and the new versions in parallel in the code for a while)
we simply duplicate the list and detail pages for alerting
configurations (as event filters) and alerting integrations (as alert
channels) for now.