Use data from any beacon
The beacon summary at the top of the page load view gets its data from
the first beacon.

For SPAs it can be the case that the first beacon has no user data, e.g.
when the page flow is as follows:

 - sign in
 - successful sign in (user data gets set)
 - ...

This change will use the user provided on any of the beacons found
within the page load.