Remove duplicate ticks (#1339)
**Why**

We reoeatedly have duplicate tick labels in charts when the formatter
does not support decimal places and when the range is pretty small.
Duplicate labels look like a bug.

**What**

Check whether formatters support decimal places. If they do, do not
manipulate the tick interval. If they don't set the minimum tick
interval to `1`.

Note that this PR is still incomplete. We must pass in the formatter via
all code paths to make sure that all charts are making use of this
capability.

**References**

 - [Ticket](https://instana.zendesk.com/agent/tickets/6237)