fix excessive, repeated http calls/loading of entities for the list
Turns out, the order matters. This places the connectTo first in the
compose call and all the withState after it.

Before the change, any change of the state would re-execute the
connectTo and thereby trigger the list request again. Now the list is
loaded only once, as it should be.