hide magnifying glass when search input has text
In Safari, the magnifying glass would overlap the little x icon that
clears the search field's content because of this:
https://github.com/instana/ui-client/blob/f0e514ffd48ffb51f336e6e9065f59cce1f488d3/packages/in-themes/foundation.less#L31-L33

Note: That x-icon to clear the content is browser chrome, we do not
render it ourselves.

This change makes the magnifying glass disappear as soon as there is
text content in the search input field. This way, we can drop
padding-right (which served to avoid that overlap). This is also
consistent with how the search field behaved before 2018/10.

An alternative would be to drop the extra rule for Safari referenced
above (the one that sets padding: 0 ! important), it seems the most
recent Safaris do not have
[this](https://github.com/zurb/foundation-sites/issues/5755)
[problem](https://bugs.webkit.org/show_bug.cgi?id=111843) anymore.

But even if that could be verified, older Safaris still have the issue,
so hiding the magnifying glass and dropping the padding-right seems to
be the safer route.