use memoized observable to prevent reloading referenced entities on every form change
Turns out the UpdateOnlyWhenChanged componet was a really stupid idea -
it kept the old/stale form around in the closure for the onSubmit
handlers in for the SelectListDialogs.

Using a memoized observable is the far better alternative to prevent
reloading the referenced entities on every change of the form values.