Fix rolling rollout
*Why*

During the rollout of ui-client we will temporarily have a new and an
old version of ui-client running. This can cause problems when the
request serving the `index.html` file is hitting the new ui-client and
the request serving the `index.js` is hitting the old ui-client.

We must ensure that ui-client is behaving consistently when serving the
`index.js` file. When the requests hits the wrong ui-client, then
ui-client must respond accordingly.

This problem gets more complicated through the in-browser integrity
check. Because of this check the whole UI will fail to render.

*What*

Browsers request `index-{checksumOfFile}.js`. ui-client can check that
the checksum matches what it could serve. When it doesn't match, it
should return a 404. The proxy can then make use of this status code to
attempt a call to another ui-client. Eventually it will hit an updated
ui-client.