Fix Auth Check
*Why*

The authentication check came too late in the request handling flow. As
a result of it being executed too late, other requests (that depend on
valid auth) were resulting in non 200 status codes.

*What*

This change puts the status code checks earlier into the flow. It also
cleans up the code slightly.

And it contains an unrelated change that was still in my stash...: Show
showing in the page load detail view instead of a broken and unstyled
error message (error message is shown when the UI is outdated and shared
component updated).