Present entity specific icon (#1554)
*Why*

Because we shouldn't use a Linux icon for a Windows or other hosts. This
is a regression which happened as part of the design rework.

*What*

Retrieve the icon for a snapshot and not just for the plugin.

Also remove all knowledge about the plugin system from the `Breadcrumb`
component to reduce coupling.