Improve Tooltip Performance (#1627)
* Improve Tooltip Performance

**Why**

While running the full backend locally I got to experience quite the
slowdown when making use of the chart's tooltips. More specifically when
hovering over them.

This is because we are updating the tooltip position in all charts on
every mouse move.

**What**

While we should eventually update the tooltip position, we do not need
to do so for every tiny mouse movement. This is just wasting CPU
resources.

* Clarify comment