Improve agent update mechanism (#1740)
**Why**

It is currently only visible in dev mode and therefore not easily
accessible in SaaS. It also doesn't match the presentation style for
other internal features.

**What**

Show the feature when internal mode is enabled. Also make use of the
same button card style that we use in the Dropwizard and agent
dashboards.