Allow dev mode against k8 environments (#1759)
**Why**

Because the old fullstack boxes are deprecated and we need to ensure
that we can continue to work.

**What**

The main challenge here are the new base domains. Most specifically
`{{env}}.instana.rocks`. This results in the cookies being restricted to
these base domains. Instead of fiddling with the backend configuration
to allow access to cookies from more domains, this PR ensures that we
can work with this tighter security restriction.

The most visible consequence is that we are working with multiple
different local domains now (see `README.md`). Everyone needs to ensure
that these are locally configured.

Upon dev mode start, the URL will be opened against the correct URL
automatically.