Remove CSRF request failure logging to instanaops (#1807)
**Why**

We log CSRF request failure errors to Instana and that is unnecessarily
increasing the error rate. This is especially curious because we changed
the logging behavior to log this on a debug level.

https://eu-instanaops.instana.io/#/websiteMonitoring/website;websiteId=854_zviATDW1EkZ9-teR2Q/errors;orderBy=errorsAgg;orderDirection=DESC;page=1;pageSize=20;query=csrf?timeline.to&timeline.ws=86400000&timeline.fm

**What**

While analyzing this I noticed that error signals aren't being correctly
forwarded to subscribers in the `.transform`, `.flatMap` and `.merge`
operators.

The CSRF implementation is actually correct. Only the
`reactive-observables` bug is stopping it from working properly.

This PR fixes the `reactive-observables` impl.

Does this mean that we won't be seeing the CSRF errors anymore in
Instana? No, Instana keeps track of request performance automatically.
What we are doing currently is duplicating that information.