Show referenced page load details in custom beacon details (#1850)
**Why**

Adidas wants to accumulate data for multiple consecutive page loads
within a single view. We discussed an option with them to use our custom
beacon mechanism for this. This gives us time to properly evaluate how
this use case can be solved.

One downside of custom beacons is that the data is really just that:
Custom. We have very little understanding what customers put into them.
This is a downside for their use case, as they want to inspect all the
referenced page loads.

**What**

Respond in a special way to referenced page load IDs in beacon meta
data. Render basic information about the page load, i.e. the URL and
duration, within a preview. This allows users to navigate to the
problematic page load.

**References**

 - [Zendesk](https://instana.zendesk.com/agent/tickets/8859)