JavaScript Stack Trace Translation Configuration (#1854)
* JavaScript Stack Trace Translation Configuration

**Why**

We want users to be able to get nicely parsed and translated stack traces. One
absolute requirement for this is that we can download the JavaScript and source
map files. In order to get passed access control we need to enable users to
configure credentials.

**What**

Support configuration of credentials within the website configuration area.

**References**

 - [Story](https://www.pivotaltracker.com/story/show/162444958)
 - [Miro](https://miro.com/app/board/o9J_kxggBos=/?utm_source=notification&utm_medium=email&utm_campaign=mentions&utm_content=header)

* Feature flag the extended configuration dialog

* Get started with the dialog

* Validations

* Add config flow

* Prepare editing

* Move website monitoring subscriptions over to in-websites

* Support removal of items and add storybook for it

* Use component from settings area for the base design

* Small design adaptations