Improve WebSocket connection handling (#1868)
* Improve WebSocket connection handling

**Why**

Sometimes users open the Instana window in the background, i.e. the
window is not visible. In these cases we will establish the
subscriptions to the backend even though the user cannot see any of the
data.

This gets resource intensive when the UI is in live mode and somewhere
within the analyze area. It gets even worse when the user has a spotty
internet connection resulting in frequent...

 1. loss of connectivity
 2. subscription establishment (aka. new subscriptions)
 3. loss of connectivity
 4. ...

**What**

We can save resources by waiting with subscription establishment until
the tab is visible. If it isn't visible, do not even establish any
subscriptions.

Furthermore it would be nice to understand when backend connections were
established and when they were lost. Unfortunately this created
complications as this introduced an import cycle:

 1. mixpanel needs subscriptions (for properties reported to Mixpanel)
 2. subscriptions need the connections
 3. connection needs Mixpanel in order to report connection events.

In order to resolve this look this PR also does the following:

 1. Continue the previous tracking refactoring by turning Mixpanel into
    just one of potentially many tracking service providers (we have two:
    Mixpanel and Instana).
 2. Establish a tracking service provider (`in-services/tracking/trackers`)
    with which the various tracking services can register. This then
    resolves the import cycle.
 3. Remove all previous `createTracker` calls so that the UI code base
    is consistent in their usage of the `track` method from the
    `in-services/tracking/tracking` module.

* Report connection type