Recursively expand nested JSON (#1900)
**Why**

Because our tracer and EUM API only allow users to define custom tag values as strings, but nested JSON
is kinda common. It would be nice to present it properly as JSON.

**What**

Recursively try to parse the nested JSON to generate something more readable in the UI.

The EUM area already did this. Ensure that both places use the same new and better logic.