Improve appdata query perf internal dashboard (#1905)
**Why**

Because the previous dashboard only worked as long as we had one
appdata-reader. We have two since a few months now. So it is completely
pointless.

Also establish references to the most common appdata analyze queries to
make repeated analysis of appdata performance easier.

**What**

Establish a new component that allows the presentation of
`getCallGroups` query in a regular chart. We already have this for
websites. So adapt it for usage in the appdata area.

Link to the analyze area to support more fine grained queries if one
wants to do this.