URL Shortener (#1916)
* URL Shortener

**Why**

Our URLs are extremely useful. They permit easy sharing of what you are seeing
with colleagues, bookmarking, proper browser history and… break disrupt
communication.

Wouldn't it be nice to support URL shortening?

**What**

Show a prompt when users press the shortcut to access the browser address bar. This
prompt asks users whether they would like to have a short version of their URL.
When they click it, then we generate a short URL and allow the users to copy it.

Obviously some users might not want it, so they can disable it with the click of a
button. If they do decide afterwards that they want it, then they can re-enable
it via the settings dialog.

* Review