Improve beacon list interaction model in page load view (#1933)
* Improve beacon list interaction model

**Why**

The beacon list currently contains a clickable element that looks like a hyperlink, but isn't one.
This is not only confusing, but customers also do not figure out easily that the whole sections are
expandable.

**What**

Make the whole beacon row clickable and properly handle keyboard and focus events. The later has been
an ongoing topic within various parts of the UI. This PR introduces a reusable function that ensures
that we implement the W3C WAI-ARIA guidelines (or at least a subset) and that we maintain standard
browser keyboard interaction patterns, e.g. tab, shift+tab, space and enter keys to interact with
interactable HTML elements.

**References**

 - [Task](https://www.pivotaltracker.com/story/show/165143461)

* Review