Cleanup metric subscriptions (#1882)
** Why
Currently, when we build infrastructure metric charts, we join the live metrics with the historic metric series on the UI. This presents a problem because the live metric subscription that is used there is also used to show a single metric or KPI. To make these KPIs responsive, in the backend we also fetch the latest value from Cassandra in "live" metric requests, because the rollups may take some time to come. This behavior actually means that when we query for a metric series for a chart, we might query Cassandra twice: 1) for the historic metric series, and 2) for the latest value (which will be included in 1). The cost of this is not very great, in fact, it's possible both these queries are being collapsed, but I am not sure about that. However, for the new ad hoc aggregation queries, this would be a problem, because each extra historic query may result in tens, hundreds, or even thousands of Cassandra queries.

Therefore, I am attempting to refactor the metric subscriptions in the backend to do a significant cleanup - which was needed anyway.

This is necessary for fixing a couple of issues with the K8s CPU and Memory aggregations.

** What
This PR converts the UI to use the new subscribe-metrics and subscribe-latest-metrics endpoints which will automatically provide live updates depending on the autoRefresh parameter.

- subscribe-metrics: returns a series of metrics between timeConfig.from and timeConfig.to with a resolution determined by the rollup and whether or not to send live updates determined by autoRefresh.
- subscribe-latest-metrics: returns a single metric value, with a time determined by timeConfig.to, a resolution determined by rollup and whether or not to send live updates determined by autoRefresh.