Handle incoming references from logging integrations (#2039)
* Handle incoming references from logging integrations

**Why**

We added links from Instana to Humio, Splunk and LogDna. Now it is
time that they link back to us. We need to resolve the payload we
are giving to useful links to entities, then present these links
in order to achieve a good experience.

**What**

Register a new dashboard that users will land on when following
links from integrations. The dashboard transmits the config
received from the integration to the backend in order to
identify what kind of entities we should link to.

The matrix parameter `config` for the URL segment `/landing`
contains URL encoded JSON that contains all the data provided
from integrations.

**Development Hints**

This is a good test URL to experiment:

https://local-instana.instana.io:4000/#/integration/landing;config=%7B"kubernetes.host.id"%3A"2131"%2C"timestamp"%3A10000%7D?timeline.to&timeline.ws=3600000

**TODOs**

 - [ ] Implement the presentation logic for the references we
       received from the backend.

**References**

 - [Required backend PR](https://github.com/instana/backend/pull/4478)
 - [Documentation](https://docs.google.com/document/d/1nLqbPw2zoa8V6OjlqxaUbFRnQV_yUTc8_O9ZQsWvylQ/edit?ts=5d9b9a06)