Display a warning message on the k8s summary dashboards (cluster, node, namespace, deployment, deploymentConfig, service, pod) if there are missing resource permissions for deployments and replicasets within the 'apps' apiGroup. This means that the agent is being deployed by a very old version of the daemonset/helm chart that does not include the "apps" cluster role for deployments & replicasets.
We want to remove all references to the old extensions/v1beta1 k8s API in the Kubernetes Sensor in order to move forward with support for k8s v1.16. So customers still using very old versions of the daemonset/helm chart need to update to the latest in order for the right permissions to be set.