"Inbound or All calls" on Application Perspective Dashboard (#2148)
# Why

Applications Perspectives is a concept that is used in several ways in Instana. Some see Applications as a black box or as a greater service, whereas others use Applications to monitor a group of services and their inner calls.

Application Perspectives should enable the user to chose between the "black box" monitoring, showing metrics relative to call coming from the outside of the Application, or to choose a "white box" monitoring, showing not only calls coming from outsite of the Application, but also calls which happen in-between its services.

# What

Added a 'matrix-parameter' `boundaryScope` which keeps track of showing just 'inbound' or 'all' calls for a Application Perspective.
A new component with radio buttons is created (based on `ExpandableCard` and `OptionBox`) where the user can switch between the two options.
The boundary scope is passed on to the `filter` for various AP queries, to filter down results and metrics.