prefer template name instead of template path (#2325)
The PHP extension is now reporting two items for Twig:

- `template path` and `template name`

The `name` is preferable because it contains a name that is context
specific to Symfony. If that is missing, we fall back to the full path
for the template on the filesystem. Due to UX reasons, we only take
the filename in that case and limit display of the full path to the
span plugin.