Fix uncaught error in infra list (#2390)
**Why**

We have an uncaught error that occurs within the infrastructure list of the
AP product area:

https://instana.io/s/CaY4yO0sTYu9a1Uz35bTLQ

**What**

Never put immutable objects into the location objects. The location objects
are mutable for the sake of efficiency. Making them immutable means that all
logic would either need to check for immutability of always make copies.

While trying to locate this bug I replaced `withUrlDependingState` with
the newer `withUrlState`. This wouldn't have been necessary in the end,
but since we want to migrate to `withUrlState` anyway, I didn't wanna lose
that change.