Add POD_IP environment variable to the instana-agent container within the DaemonSet to make binding more stable.
The agent by default binds to loopback addresses and what it perceives to be public IPs as well as docker bridge addresses. In some cases, it will not bind to the pod IP, which will cause the liveness checks to fail.

The agent looks for the POD_IP environment variable and will bind to that address instead of trying to detect the public IP address.
This PR adds this environment variable so that the agent can always bind to the right address. This should be more stable in the presence of different network plugins or configurations.

[This PR is the originator of the change](https://github.com/instana/instana-agent-charts/pull/5)

And see this Slack discussion in #product-feedback as to why the Instana Agent YAML is in the ui-client: https://instana.slack.com/archives/C04FG7TFS/p1576520790008200