Fix syntax error in Edge
**Why**

Syntax errors occur in Microsoft Edge because it does not yet
support the object spread operator natively. When we moved to
`@babel/present-env` we accidentally removed the object-spread
transformation (because we aren't using object-rest) from the
build process. This caused the syntax:

```
{
  ...foobar
}
```

to no longer be transpiled.

**What**

Re-add the transformation to the build process.