Remove Google Analytics (#2495)
**Why**

For some time now we had Google Analytics disabled for the product because
we believed that nobody is looking at this data. This allowed us to get
better at *Datensparsamkeit*.

It was now brought up to product managements attention that this data
was not missed by anybody.

**What**

Remove all Google Analytics code now.