Get insights into script errors (#2555)
**Why**

Because we frequently have script errors in SaaS and we have no
idea where they are coming from.

**What**

Add `crossorigin=anonymous` to third-party script tags. Unfortunately
this will not allow us to get insights into all third-party scripts,
but maybe we get lucky and the errors are already in those that are
now annotated with `crossorigin=anonymous`?