show only upstream or downstream in flowmap (#2695)
# Why & What

Upstream/Downstream context guide provides a link to see all upstream/downstream services/endpoints. The link opens the flowmap of the current service/endpoint and all the upstream and downstream entities are expanded.

It'd be better to only expand upstream or downstream entities depending on which tab we come from.