Do not retrieve event chart when invisible (#2732)
**Why**

We know that the event count over time retrieval is quite expensive
for customers with large event indices. It is therefore desirable to
avoid execution of the query whenever possible.

**What**

Luckily the event chart is located at the end of the new system
overview dashboard. This enables us to delay retrieval of the event
count over time until users actually scroll down to the chart.

Unfortunately there is one small complication with react-grid-layout's
usage of transitions which is outlined within a JavaScript comment.