Finish dashboard switcher and integrate into Cockpit (#2776)
**Why**

The custom dashboard list shall not exist as per the design. Instead
we shall have a dropdown within the header that can be used to switch
between custom dashboards.

**What**

Build the new component and hook it up in the Cockpit (when the FF is
enabled). Unfortunately it cannot be used yet within the new custom
dashboard / custom dashboard edit flow, because the whole mode is
supposed to work different than anticipated.

Follow up PRs will change the new/edit mode and then the switcher can
be integrated.

The PR furthermore discovered a bug in our list component that is now
fixed as well: The `Li` creates new React components within every
rendering step hence breaking all change detection and state management.

**References**

 - [Design](https://miro.com/app/board/o9J_kvPsifc=/)