New dashboard creation flow (#2778)
**Why**

Within the design a dashboard is created by first typing in a name and then
waiting for the save action to be finished. This is beneficial for the
following dashboard edit / sharing mode behavior (it makes it easier).

Within that design the `Create new dashboard` button in the dashboard switcher
triggers this dialog.

**What**

Create a new dialog according to the design. While doing so I noticed that the
`BigHeaderDialog` has certain assumptions around its content, e.g. it requires
the usage of a `SlideInView`. This PR lifts this restriction.

Furthermore we do not seem to have standardized save and cancel actions. I
therefore created components for this under our standard form package. They
implement best practices for our forms, e.g. waiting animations.

**References**

 - [Design](https://miro.com/app/board/o9J_kvPsifc=/)