Add a badge with a tooltip explaining total latency of batched calls. (#2782)
# Why

In UA latency of batched calls is shown as total latency of all calls in the batch. This could be confusing when filtering by latency, e.g. `latency <= 49ms`, since the total latency of batch calls could be higher than the desired max latency by the multiple of the batch size.   

# What

To avoid this confusion, a badge with a tooltip _"Total latency of X batched calls."_  will be shown next to the latency of a batched call.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/11988/details/