Introduce an expandable menu (#2812)
**Why**

Introduce a standardized component for expandable menus. We are starting to
see more of these pop up and we should avoid siloed implementations.

The need for a standardized one comes from the custom dashboard work in which
a dashboard header needs to contain such an expandable menu.

**What**

 - Introduce the standardized component.
 - Use the component in the custom dashboard header.

**Scope**

This PR deliberately does not update other locations in the code in which
we have separate expandable menu implementations. To my knowledge these are:

 - AP dependency map
 - Charts