Custom Dashboard Sharing Dialog (#2821)
**Why**

We need to provide the ability to share who can access a custom dashboard.
Access is further differentiated between read and write access.

**What**

Add a sharing dialog as shown in the design, but only implement a subset,
i.e. granting global read access. Granting per-user write access is
currently not implemented as I am waiting for feedback on required
API changes.

In order to implement the design I had to make small changes to `CheckboxFancy`.
The checkbox/radio button now supports explanatory texts in order to be
able to implement the design. I can imagine that this one off design will
become standardized in the future, but this isn't completely thought through
in design. If it does become standardized, we can make the introduced `Option`
component reusable.

**References**

 - [Design](https://miro.com/app/board/o9J_kvPsifc=/)
 - [API change discussion](https://instana.slack.com/archives/GBV0833SN/p1584439245013100)