Introduce light cards to custom dashboards (#2866)
* Introduce light cards to custom dashboards

**Why**

We want to make use of the light cards in more and more product areas.
I specifically had a need for them within the custom dashboard area.

**What**

This PR introduces a standardized mechanism to switch between the two
*themes* we are currently supporting:

 - The current light mode (`light`)
 - The next generation light mode (`lightV2`)

By default `light` is enabled globally. For the time being `lightV2` can
be enabled on a cases by case basis via `LocallyChangedTheme`. The `Card`
component observes the theme (via React contexts) and updates its
presentation style automatically.

In the future we will ideally get rid of `light` and re-label `lightV2`
to `light`. Then we will hopefully introduce dark mode as well.

**References**

 - [Design](https://miro.com/app/board/o9J_kvPsifc=/)

* Fix light card v2 behavior when using max available width combined with content that would normally not consume the available vertical space

* Also support light mode for KPIs

* Sizing behavior changes

* Fix widget sizing so that end-users have more control over the size of each widget

* Adjust minimum sizes of custom dashboard widgets